+++ src/backend/mod.rs
@@ -1,4 +1,4 @@
-#[cfg(feature = "freerdp")]
+#[cfg(tiny_shell_freerdp_backend)]
 pub(crate) mod freerdp;
 pub(crate) mod local;
 pub(crate) mod remote_desktop;